split character controller and collider
